moving multilabel functions
