+++ cleanlab/internal/multilabel_utils.py
@@ -20,7 +20,7 @@ Helper functions used internally for multi-label classification tasks.
 
 import numpy as np
 
-from cleanlab.internal.util import get_num_classes, int2onehot
+from cleanlab.internal.util import get_num_classes
 
 
 def _is_multilabel(y: np.ndarray) -> bool:
@@ -64,3 +64,35 @@ def get_onehot_num_classes(labels, pred_probs=None):
             "wrong format for labels, should be a list of list[indices], please check the documentation in find_label_issues for further information"
         )
     return y_one, num_classes
+
+
+def int2onehot(labels) -> np.ndarray:
+    """Convert list of lists to a onehot matrix for multi-labels
+
+    Parameters
+    ----------
+    labels: list of lists of integers
+      e.g. [[0,1], [3], [1,2,3], [1], [2]]
+      All integers from 0,1,...,K-1 must be represented."""
+
+    from sklearn.preprocessing import MultiLabelBinarizer
+
+    mlb = MultiLabelBinarizer()
+    return mlb.fit_transform(labels)
+
+
+def onehot2int(onehot_matrix) -> list:
+    """Convert a onehot matrix for multi-labels to a list of lists of ints
+
+    Parameters
+    ----------
+    onehot_matrix: 2D np.ndarray of 0s and 1s
+      A one hot encoded matrix representation of multi-labels.
+
+    Returns
+    -------
+    labels: list of lists of integers
+      e.g. [[0,1], [3], [1,2,3], [1], [2]]
+      All integers from 0,1,...,K-1 must be represented."""
+
+    return [list(np.where(row == 1)[0]) for row in onehot_matrix]
