+++ tests/test_filter_count.py
@@ -19,7 +19,8 @@ from cleanlab.count import get_confident_thresholds
 from cleanlab.internal.latent_algebra import compute_inv_noise_matrix
 from cleanlab.benchmarking.noise_generation import generate_noise_matrix_from_trace
 from cleanlab.benchmarking.noise_generation import generate_noisy_labels
-from cleanlab.internal.util import value_counts, int2onehot
+from cleanlab.internal.util import value_counts
+from cleanlab.internal.multilabel_utils import int2onehot
 import numpy as np
 import scipy
 import pytest
